+++ demo/components/Menu/index.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { Story } from '../Story';
+import { BasicUsage } from './BasicUsage';
+
+export const Menu = (): React.ReactElement => (
+  <Story title="Menu">
+    <BasicUsage title="Basic example" />
+  </Story>
+);