+++ demo/cypress/index.ts
@@ -17,6 +17,7 @@ export { FileUpload } from './FileUpload';
 export { Input } from './Input';
 export { Loader } from './Loader';
 export { MaskedInput } from './MaskedInput';
+export { Menu } from './Menu';
 export { Modal } from './Modal';
 export { MultiSelect } from './MultiSelect';
 export { Notifications } from './Notifications';