+++ cypress/integration/Modal/index.spec.js
@@ -1,61 +1,94 @@
+import { defaultModalTheme as theme } from '../../../korus-ui/components/Modal/theme'
+
 /* eslint-disable no-unused-expressions,jest/valid-expect */
 describe('Modal', () => {
   before(() => {
     cy.visit('/cypress/modal');
   });
 
-  describe('Open and Close', () => {
-    it('should be closed', () => {
-      cy.contains('Modal header')
-        .should('not.be.visible')
-    });
-
-    it('should be open', () => {
-      cy.get('[data-test=openModalButton]').click()
-      cy.contains('Modal header')
+  describe('Interaction', () => {
+    it('Should open and close', () => {
+      cy.datatest('openModalButton')
+        .click()
+        .get(`.${theme.window}`)
         .should('be.visible')
+        .get(`.${theme.cross}`)
+        .click()
+        .get(`.${theme.window}`)
+        .should('not.exist')
     });
 
-    it('should be closed', () => {
-      cy.get('.modal-window > .modal-cross').click()
-      cy.contains('Modal header')
-        .should('not.be.visible')
-    });
-  });
-
-  describe('Alert', () => {
-    it('should open Alert', () => {
-      cy.get('[data-test=openModalButton]').click()
-      cy.contains('Simple alert')
-        .should('not.be.visible')
-      cy.get('[data-test=openAlertSimpleButton]').click()
-      cy.contains('Simple alert')
+    it('Closing by close button', () => {
+      cy.datatest('openModalButton')
+        .click()
+        .get(`.${theme.window}`)
         .should('be.visible')
+        .get(`.${theme.cross}`)
+        .click()
+        .get(`.${theme.window}`)
+        .should('not.exist');
     });
 
-    it('should hide Modal close button', () => {
-      cy.get('.modal-window > .modal-cross')
-        .should('not.be.visible')
-    });
-
-    it('should disable Modal close on overlay click', () => {
-      cy.get('.modal-wrapper').click()
-      cy.contains('Simple alert')
+    it('Closing by cancel button', () => {
+      cy.datatest('openModalButton')
+        .click()
+        .get(`.${theme.window}`)
         .should('be.visible')
+        .get('#cancel')
+        .click()
+        .get(`.${theme.window}`)
+        .should('not.exist');
     });
 
-    it('should close Alert', () => {
-      cy.get('.modal-alert > .modal-alert-cross').click()
-      cy.contains('Simple alert')
-        .should('not.be.visible')
-    });
-
-    it('should close Modal', () => {
-      cy.contains('Modal header')
+    it('Closing by click outside the modal window', () => {
+      cy.datatest('openModalButton')
+        .click()
+        .get(`.${theme.window}`)
         .should('be.visible')
-      cy.get('.modal-window > .modal-cross').click()
-      cy.contains('Modal header')
-        .should('not.be.visible')
+        .get(`.${theme.wrapper}`)
+        .click(0, 100)
+        .get(`.${theme.window}`)
+        .should('not.exist');
+    });
+
+    it('Closing with the ESС key', () => {
+      cy.datatest('openModalButton')
+        .click()
+        .get(`.${theme.window}`)
+        .should('be.visible');
+      cy.focused().type('{esc}');
+      cy.get(`.${theme.window}`)
+        .should('not.exist');
+    });
+
+    it('Leave alert', () => {
+      cy.datatest('openModalButton')
+        .click()
+        .datatest('openAlertLeaveButton')
+        .click()
+        .get('#leave')
+        .should('be.visible')
+        .get('#okButton')
+        .click()
+        .get('#leave')
+        .should('not.exist')
+        .get('#cancel')
+        .click();
+    });
+
+    it('Simple alert', () => {
+      cy.datatest('openModalButton')
+        .click()
+        .datatest('openAlertSimpleButton')
+        .click()
+        .get('#simple')
+        .should('be.visible')
+        .get('.modal-alert-cross')
+        .click()
+        .get('#simple')
+        .should('not.exist')
+        .get('#cancel')
+        .click();
     });
   });
 });