+++ storybook/storybook/stories/form/Input/Examples/index.js
@@ -2,8 +2,10 @@ import componentSrc from '!raw-loader!@korus/leda/dist/components/Input/';
 
 export { componentSrc };
 export { AllowedSymbols } from './AllowedSymbols';
+export { Api } from './Api';
 export { BasicUsage } from './BasicUsage';
 export { Customization } from './Customization';
+export { ForbiddenSymbols } from './ForbiddenSymbols';
+export { HasClearButton } from './HasClearButton';
 export { LetterCase } from './LetterCase';
 export { MaxLength } from './MaxLength';
-export { Api } from './Api';