+++ storybook/storybook/stories/form/NumericTextBox/Examples/index.js
@@ -1,3 +1,7 @@
 export { Api } from './Api';
 export { Customization } from './Customization';
 export { Formatting } from './Formatting';
+export { Limits } from './Limits';
+export { ThousandsSeparator } from './ThousandsSeparator';
+export { TrailingZeros } from './TrailingZeros'
+