+++ storybook/storybook/stories/form/NumericTextBox/NumericTextBox.stories.js
@@ -5,6 +5,9 @@ import
   Api,
   Customization,
   Formatting,
+  Limits,
+  TrailingZeros,
+  ThousandsSeparator,
 } from './Examples';
 import { Story } from '../../../components/Story';
 import { propsDesc } from './propsDescription';
@@ -12,7 +15,10 @@ import { ApiStory } from '../../../components/ApiStory';
 
 storiesOf('Form|NumericTextBox', module)
   .add('API', context => (<ApiStory {...Api} customProps={propsDesc} context={context} />))
-  .add('Customization', context => (<Story {...Customization} customProps={propsDesc} context={context} />));
+  .add('Кастомизация', context => (<Story {...Customization} customProps={propsDesc} context={context} />))
+  .add('Ограничение диапазона', context => (<Story {...Limits} customProps={propsDesc} context={context} />));
 
 storiesOf('Form|NumericTextBox.Props', module)
-  .add('format', context => (<Story {...Formatting} customProps={propsDesc} context={context} />));
+  .add('format', context => (<Story {...Formatting} customProps={propsDesc} context={context} />))
+  .add('trailingZeros', context => (<Story {...TrailingZeros} customProps={propsDesc} context={context} />))
+  .add('thousandsSeparator', context => (<Story {...ThousandsSeparator} customProps={propsDesc} context={context} />));