+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Menu: Added new component
+- Menu: Close dropdown items on scroll
 
 ### Changed
 - Modal: Prop 'size' adds class to div.modal-window instead of style attribute