+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.42.0] - 2021-02-25
+## [0.42.0] - 2021-03-15
 
 ### Added
 - CountDown: Added new component
+
+### Fixed
+- Tabs: Fixed scroll
+
 
 ## [0.41.0] - 2021-02-25
 
@@ -20,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - DatePicker: Fixed date selection from previous month
+
 
 ## [0.40.0] - 2021-02-09